Addresses #26 - Add preliminary tests, do not optimize destructively

+++ src/zimscraperlib/image/optimization.py
@@ -3,6 +3,8 @@
 # vim: ai ts=4 sts=4 et sw=4 nu
 
 import shutil
+import tempfile
+import pathlib
 
 from optimize_images.data_structures import Task
 from optimize_images.do_optimization import do_optimization
@@ -11,66 +13,82 @@ from .. import logger
 
 
 class ImageOptimizer:
-    def __init__(self, png_options=None, jpeg_options=None):
+    def __init__(self, png_options={}, jpeg_options={}):
         self.png_options = png_options
         self.jpeg_options = jpeg_options
 
-    def optimize_png_jpeg(self, src, dst, format, override_options={}):
+    def optimize_png_jpg(self, src, dst, image_format, override_options={}):
+        # use a temporary file as source as optimization is done destructively
+        tmp_fh = tempfile.NamedTemporaryFile(delete=False, suffix=src.suffix)
+        tmp_fh.close()
+        tmp_path = pathlib.Path(tmp_fh.name)
+        shutil.copy(src, tmp_path)
+
         # generate task for optimize_images
-        if format == "jpeg":
+
+        # max_w and max_h is 0 because we have a better image resizing function in scraperlib already
+        # remove_transparency, reduce_colors, max_colors, convert_all, conv_big, bg_color, force_del, are specific to PNG
+        # quality, keep_exif are specific to JPEG
+        if image_format == "jpg":
             # generate options for jpeg
             optimization_task = Task(
-                str(src.resolve()),
-                override_options.get("quality", self.jpeg_options.get("quality", 90)),
-                False,
-                False,
-                255,
-                0,
-                0,
-                override_options.get(
+                src_path=str(tmp_path.resolve()),
+                quality=override_options.get(
+                    "quality", self.jpeg_options.get("quality", 50)
+                ),
+                remove_transparency=False,
+                reduce_colors=False,
+                max_colors=256,
+                max_w=0,
+                max_h=0,
+                keep_exif=override_options.get(
                     "keep_exif", self.jpeg_options.get("keep_exif", False)
                 ),
-                False,
-                False,
-                False,
-                (255, 255, 255),
+                convert_all=False,
+                conv_big=False,
+                force_del=False,
+                bg_color=(255, 255, 255),
                 grayscale=override_options.get(
                     "grayscale", self.jpeg_options.get("grayscale", False)
                 ),
-                ignore_size_comparison=True,
+                no_size_comparison=True,
                 fast_mode=override_options.get(
                     "fast_mode", self.jpeg_options.get("fast_mode", False)
                 ),
             )
         else:
             # generate options for png
+
+            # convert_all converts all PNG to JPEG, hence set to False
+            # conv_big converts big PNG images to JPEG, hence set to False
+            # force_del deletes the original PNG after convertion to JPEG if the above two options are used, so kept False
             optimization_task = Task(
-                str(src.resolve()),
-                90,
-                override_options.get(
+                src_path=str(tmp_path.resolve()),
+                quality=90,
+                remove_transparency=override_options.get(
                     "remove_transparency",
                     self.png_options.get("remove_transparency", False),
                 ),
-                override_options.get(
-                    "reduce_colors", self.png_options.get("reduce_colors", False)
+                reduce_colors=override_options.get(
+                    "reduce_colors", self.png_options.get("reduce_colors", True)
                 ),
-                override_options.get(
-                    "max_colors", self.png_options.get("max_colors", 255)
+                max_colors=override_options.get(
+                    "max_colors", self.png_options.get("max_colors", 256)
                 ),
-                0,
-                0,
-                False,
-                False,
-                False,
-                False,
-                override_options.get(
+                max_w=0,
+                max_h=0,
+                keep_exif=False,
+                convert_all=False,
+                conv_big=False,
+                force_del=False,
+                bg_color=override_options.get(
                     "background_color",
                     self.png_options.get("background_color", (255, 255, 255)),
                 ),
                 grayscale=override_options.get(
                     "grayscale", self.png_options.get("grayscale", False)
                 ),
-                ignore_size_comparison=True,
+                no_size_comparison=True,
                 fast_mode=override_options.get(
                     "fast_mode", self.png_options.get("fast_mode", False)
                 ),
@@ -81,22 +99,26 @@ class ImageOptimizer:
         if result.was_optimized:
             logger.info(f"Successfully optimized image: {src}")
             logger.debug(f"Space savings: {result.orig_size - result.final_size}")
-        else:
-            logger.warning(f"{src} not optimized")
-        if src.resolve() != dst.resolve():
-            logger.info(f"Copying {src} to {dst}")
-            shutil.copy2(src, dst)
+            shutil.copy(tmp_path, dst)
+            tmp_path.unlink()
+        return result.was_optimized
 
-    def optimize_image(self, src, dst, delete_src=True):
+    def optimize_image(self, src, dst, delete_src=True, override_options={}):
         # optimize the image with the correct optimizer
         if not src.is_file():
             raise FileNotFoundError("The requested image is not present")
+        optimized = False
         if src.suffix in [".jpeg", ".jpg"]:
-            self.optimize_png_jpeg(src, dst, format="jpeg")
+            optimized = self.optimize_png_jpg(
+                src, dst, image_format="jpg", override_options=override_options
+            )
         elif src.suffix == ".png":
-            self.optimize_png_jpeg(src, dst, format="png")
+            optimized = self.optimize_png_jpg(
+                src, dst, image_format="png", override_options=override_options
+            )
         else:
-            logger.error("Image format not supported for optimization")
-            return
-        if src.resolve() != dst.resolve() and src.exists() and delete_src:
+            raise Exception("File not supported for optimization as an image")
+
+        # delete src image if requested
+        if src.resolve() != dst.resolve() and src.exists() and delete_src and optimized:
             src.unlink()
